Simplify VCR property handling.
Properties like seed, type, magic are now treated as auxiliary information
that is given from a `game::vcr::Battle` to a script consumer.  This reduces
the ad-hockery needed to implement VCR properties for the script interface
and VcrPacker.

As a result, all combat now has `Loc.X`, `Loc.Y` properties, and FLAK
properties are published in a straightforward way.