Simulator: implement battle order peculiarities
In PHost, aggression is not symmetric. We therefore need to track who is
aggressor and who is opponent, and only check the respective conditions.

In Host, aggression is mostly symmetric, but a captured ship cannot appear
on the left side (inner loop of battle order, what would be the "opponent"
role in PHost).

Fixes #428.