Graphics: add basic 3D engine
The structure is modeled after the one used in the JavaScript FLAK player.
This engine is not intended to be fully general, but only sufficient to
implement things similar to FLAK.

Included is a simple "lo-fi" software based renderer, but I assume this
interface to be implementable using OpenGL as well.

Ported from PCC2 Web:
- WebGL/LineRenderer
- WebGL/ParticleRenderer
- WebGL/TriangleRenderer