Portability: do not use covariant return types for PropertyAccessor
The toolchain (still) used to build the Windows binaries miscompiles the
pattern

    struct B { int a; };
    struct A { virtual B* get() = 0; };
    struct AB : A, B { virtual AB* get() { return 0; } };

by failing to (not) adjust the returned null pointer. A caller of `get` will
see the returned pointer as non-null, causing all sorts of code to crash.

Work around that by not using covariant return types for `lookup` until the
toolchain can be replaced.