Allow saving to trigger an asynchronous/waiting operation
`Session::save` and related functions are now based on a task that can be
active for a while. When run under script control, this will suspend the
invoking process until the task completes. In case the save operation
requires some sort of user interaction, it can suspend to execute that.

A save operation run outside script control would need some way to hold and
re-activate the task which is not currently implemented.