RequestThread: ensure correct destruction order even when discarding tasks
Tasks might reference temporaries (RequestSender::makeTemporary) that refer
to each other, so destroying them in the wrong order means a task referring
to the temporary overtakes one that destroys it.

This can occur during shutdown, when tasks that detect that they're
discarded destroy the object whose lifetime they control as a last measure.