Add support for object labels
Object labels are computed from ship/planet content using an expression, and
can be displayed in the starchart. This implements the label logic for c2ng,
in a slightly different fashion than in PCC2 (without the need for temporary
processes).

Labels are an optional Session extra and therefore only present when
instantiated for the program (i.e. not present in c2script).

Ported from PCC2:
- client/labels.cc