Move map configuration from universe to game
Map configuration remains constant within a game across turns, so storing
this information in a Universe leads to trouble when changing between turns.
It could actually even be moved into Root because it is driven by Host/User
configuration, but keeping it closer to game runtime makes more sense for
now.