Interpreter: remove unneeded typechecks after binary operations
For example, if a comparison is used as part of a complex boolean expression
such as in `x := not not (a=b)` or `x:=(a=b) or c`, a sequence such as
`bcmpeq; ubool` will be generated. The `ubool` is not needed; we know the
result to be of type boolean.