UI: rework layout computation
Remove the "minimum size" parameter. Widgets are given their preferred size
if possible. If that is not available, widgets are given less space.
Previously, the "minimum size" parameter would have given another size
limit, but in case of tight space, that would have had to be ignored as
well. Only few widgets use the ability to report different preferred and
minimum sizes, for little gain.

Rework the AxisLayout class, and clean up.