Talk service: fix message renderer bugs
Avoid that "\r" from BBCode input leaks into HTML/text output.

In text output, separate paragraphs by space. In HTML output, this problem
was previously concealed by the "\r" characters, which HTML treats as space.

When generating an abstract, make sure that we do not produce a link with
empty content that would unintentionally be rendered as link URL.