Add special "game:" file name to access game directory
`Session` now provides a special FileSystem adaptor that intercepts names
starting with "game:" and redirects those to the game directory.

This means, compiled and interpreted code using "game:XYZ" will access a
file "XYZ" in the game directory, even if that is a virtual directory.

The `System.GameDirectory` property will therefore report "game:" for
virtual directories instead of EMPTY like before. For real directories, it
keeps reporting the actual name.

This makes it possible for scripts to save data in the game directory, even
if that is on the PlanetsCentral server. Scripts that use
`System.GameDirectory` as intended do not need to be changed.