Talk Server: cross-posting
Using `POSTNEW ... ALSO x`, a thread can be cross-posted to secondary
forums. Note that a thread can only be cross-posted as a whole; all
individual postings only appear in the thread in the master forum.

Permission to cross-post is defined by `allowxpost` (unrestricted
cross-posting) and `allowgpost` (cross-post between a user's games' forums)
user permissions. The `USERLSCROSS` command can produce that list of game
forums.

With this change, the ":Seq" field in `NNTPPOSTHEAD` becomes meaningless;
NNTP server changed to no longer use it.