Prevent auto-tasks from getting lost accidentally
When preparing an auto task, and switching to another object, it might
happen that the auto tasks runs to completion (e.g. because cursor
accidentally was at the wrong place), losing the auto task.

We now add a command `CC$AutoSalvage` to the end of the task, keeping it
alive for another turn.

This feature can be turned off by setting `Task.Salvage=0` in `pcc2.ini`.