Add `c2reshack` utility
This is a port of the reshack utility included with PCC2. It has been
slightly cleaned up, but otherwise is close to the original. In particular,
it does not implement a UI-vs.-I/O split as the main client does, and is
therefore classified as a developer utility, not an end-user utility.